Refactored,renamed and update authentication service implementation
